+++ crypto/src/commonMain/kotlin/at/asitplus/awesn1/crypto/pki/X509TbsCertificate.kt
@@ -4,8 +4,10 @@
 package at.asitplus.awesn1.crypto.pki
 
 import at.asitplus.awesn1.Asn1BitString
+import at.asitplus.awesn1.Asn1BitStringish
 import at.asitplus.awesn1.Asn1Integer
 import at.asitplus.awesn1.Asn1Time
+import at.asitplus.awesn1.crypto.CursedBitString
 import at.asitplus.awesn1.crypto.SubjectPublicKeyInfo
 import at.asitplus.awesn1.crypto.Versioned
 import at.asitplus.awesn1.crypto.X509AlgorithmIdentifier
@@ -84,9 +86,9 @@ data class X509TbsCertificate(
     val subjectName: List<X500RelativeDistinguishedName>,
     val subjectPublicKeyInfo: SubjectPublicKeyInfo,
     @Asn1Tag(tagNumber = 1u)
-    val issuerUniqueID: Asn1BitString? = null,
+    val issuerUniqueID: CursedBitString? = null,
     @Asn1Tag(tagNumber = 2u)
-    val subjectUniqueID: Asn1BitString? = null,
+    val subjectUniqueID: CursedBitString? = null,
     @Asn1Tag(tagNumber = 3u)
     private val taggedExtensions: ExplicitlyTagged<List<X509CertificateExtension>>? = null,
 ) : Versioned {
@@ -110,8 +112,8 @@ data class X509TbsCertificate(
         validity = Validity(validFrom, validUntil),
         subjectName = subjectName,
         subjectPublicKeyInfo = subjectPublicKeyInfo,
-        issuerUniqueID = issuerUniqueID,
-        subjectUniqueID = subjectUniqueID,
+        issuerUniqueID = issuerUniqueID?.let { CursedBitString(it) },
+        subjectUniqueID = subjectUniqueID?.let { CursedBitString(it) },
         taggedExtensions = extensions?.takeIf { it.isNotEmpty() }?.let(::ExplicitlyTagged),
     )